+++ source/test/F0.Testing.Tests/Shared/BaseException.cs
@@ -1,0 +1,11 @@
+﻿using System;
+
+namespace F0.Tests.Shared
+{
+	public class BaseException : Exception
+	{
+		public BaseException(string message)
+			: base(message)
+		{ }
+	}
+}