+++ source/test/F0.Testing.Tests/Shared/DerivedException.cs
@@ -1,0 +1,9 @@
+﻿namespace F0.Tests.Shared
+{
+	public class DerivedException : BaseException
+	{
+		public DerivedException(string message)
+			: base(message)
+		{ }
+	}
+}