+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 CHANGELOG
 
 ## vNext
+- Added assertion for an exception to surface _synchronously_, when retrieving an asynchronous operation.
+- Added assertion for an exception to surface _asynchronously_, when awaiting an asynchronous operation.
 
 ## v0.1.0 (2019-08-31)
 - Added assertion for an exception to surface _immediately_, when retrieving an iterator.